[Feat] shakeCamera implemented and tested
Implement shakeCamera function
Implement singleShakeCamera function

Tested two functions in CameraTest.cs
