+++ to_the_STAR/Assets/Game/Scripts/CameraTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 54e90d9c71150c549b69969156d1c2c3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
