+++ to_the_STAR/Assets/Game/Scripts/ShakeCamera.cs.meta
